Fixed missing permissions on current boot for /var/run. Added user for non-package install. Fixed non-package install to work as expected.

+++ manifests/init.pp
@@ -8,13 +8,26 @@ class btsync( $webui = 'local' ) {
 
     exec { 'download-btsync':
       command => '/usr/bin/curl -o /tmp/btsync.tar.gz https://download-cdn.getsync.com/stable/linux-x64/BitTorrent-Sync_x64.tar.gz',
+      unless  => '/usr/bin/test -f /usr/bin/btsync',
     }~>
     exec { 'extract-btsync':
-      command => 'cd /usr/bin; tar xf /tmp/btsync.tar.gz btsync',
+      command     => 'tar xf /tmp/btsync.tar.gz btsync; chmod 755 /usr/bin/btsync',
+      cwd         => '/usr/bin',
+      refreshonly => true,
     }
 
     file { '/etc/systemd/system/btsync.service':
       source => 'puppet:///modules/btsync/btsync.service',
+      notify => Exec['systemd-daemon-reload'],
+    }
+
+    user { 'btsync':
+      ensure => present,
+      home   => '/var/lib/btsync',
+    }~>
+    file { '/var/lib/btsync':
+      owner => 'btsync',
+      group => 'btsync',
     }
   }
 
@@ -27,6 +40,12 @@ class btsync( $webui = 'local' ) {
     '/etc/tmpfiles.d/btsync.conf':
       ensure => file,
       source => 'puppet:///modules/btsync/btsync.conf';
+
+    '/var/run/btsync':
+      ensure => directory,
+      owner  => 'btsync',
+      group  => 'btsync',
+      mode   => '0777';
   }
 
 }
